mk-cloud-function/requirements.txt
google-cloud-spanner==3.36.0
functions-framework==3.3.0
mysql-connector-python
cloud-sql-python-connector[pymysql]
SQLAlchemy==2.0.7